Align sum result type with VTL Number Result type.
Official Sum Result type is always number (aggregate and analytic); Integer measures promote. Prefer the manual over conflicting TCK Integer fixtures, and document the Method rule in the roadmap.

+++ vtl-engine/src/test/java/fr/insee/vtl/engine/semantics/aggregation/AggregationCollectorsTest.java
@@ -33,7 +33,7 @@ class AggregationCollectorsTest {
         AggregationCollectors.fromAggrDatasetInvocation(ctx, datasetExpression, fromContext(ctx));
 
     assertThat(collectors).containsOnlyKeys("me_1", "me_2");
-    assertThat(collectors.get("me_1").getType()).isEqualTo(Long.class);
+    assertThat(collectors.get("me_1").getType()).isEqualTo(Double.class);
     assertThat(collectors.get("me_2").getType()).isEqualTo(Double.class);
   }
